Predefined types

Microfloats defines and exports a set of common types.

IEEE-like

These types have IEEE 754-like Inf/NaN encodings, with Inf being represented as all 1s in the exponent and a significand of zero, and NaN being represented as all 1s in the exponent and a non-zero significand.

Microfloats.Float8_E5M2 — Type

Float8_E5M2

Properties

Bits: 1 sign + 5 exponent + 2 significand (8 total)

Non-finite behavior: Microfloats.IEEE

Has Inf: true

Has NaN: true

Max normal: 57344.0

Min normal: 6.103515625e-5

Min subnormal: 1.52587890625e-5

Microfloats.Float8_E4M3 — Type

Float8_E4M3

Properties

Bits: 1 sign + 4 exponent + 3 significand (8 total)

Non-finite behavior: Microfloats.IEEE

Has Inf: true

Has NaN: true

Max normal: 240.0

Min normal: 0.015625

Min subnormal: 0.001953125

Microfloats.Float8_E3M4 — Type

Float8_E3M4

Properties

Bits: 1 sign + 3 exponent + 4 significand (8 total)

Non-finite behavior: Microfloats.IEEE

Has Inf: true

Has NaN: true

Max normal: 15.5

Min normal: 0.25

Min subnormal: 0.015625

Finite

These types have no Inf encoding, with alternate or no NaN encodings at all.

Microfloats.Float8_E4M3FN — Type

Float8_E4M3FN

Properties

Bits: 1 sign + 4 exponent + 3 significand (8 total)

Non-finite behavior: Microfloats.NanOnlyAllOnes

Has Inf: false

Has NaN: true

Max normal: 448.0

Min normal: 0.015625

Min subnormal: 0.001953125

Microfloats.Float8_E8M0FNU — Type

Float8_E8M0FNU

Properties

Bits: 0 sign + 8 exponent + 0 significand (8 total)

Non-finite behavior: Microfloats.NanOnlyAllOnes

Has Inf: false

Has NaN: true

Max normal: 1.7014118346046923e38

Min normal: 5.877471754111438e-39

Min subnormal: N/A

Microfloats.Float6_E3M2FN — Type

Float6_E3M2FN

Properties

Bits: 1 sign + 3 exponent + 2 significand (6 total)

Non-finite behavior: Microfloats.FiniteOnly

Has Inf: false

Has NaN: false

Max normal: 28.0

Min normal: 0.25

Min subnormal: 0.0625

Microfloats.Float6_E2M3FN — Type

Float6_E2M3FN

Properties

Bits: 1 sign + 2 exponent + 3 significand (6 total)

Non-finite behavior: Microfloats.FiniteOnly

Has Inf: false

Has NaN: false

Max normal: 7.5

Min normal: 1.0

Min subnormal: 0.125

Microfloats.Float4_E2M1FN — Type

Float4_E2M1FN

Properties

Bits: 1 sign + 2 exponent + 1 significand (4 total)

Non-finite behavior: Microfloats.FiniteOnly

Has Inf: false

Has NaN: false

Max normal: 6.0

Min normal: 1.0

Min subnormal: 0.5